Much improved variable expansion
With this change variables can be used only in part of a rule value, for
example "padding: $top $left;" will now get properly expanded.

This still doesn't handle the case of `rgba(255, 255, 255, $foo)`, as
the value is being split on spaces, but let's address that later when we
get a failing spec with the case.