Remove trailing newline inconsistencies.
This is a temporary fix so that we don't get test failures for trailing
new line where it's not supposed to be. We can figure out a more correct
way to do this later, as currently I don't really understand the rules
for trailing newlintes at the end of a file.